Use 32-bit values instead of 64 - if a sensor really needs more that 2^32 options for a single reading value they can do a custom aggregation of two streams or something.  In general most things shouldn't need more than 2^32 precision (if even that)

+++ momo_modules/mainboard/src/mib/commands/sensor_mib_feature.c
@@ -10,7 +10,7 @@ static void log_sensor_event_mib(void)
 {
 	uint8 mib_addr = plist_get_int8(0); 
 	uint8 metadata = plist_get_int8(1);
-	uint64 *value = (uint64*)plist_get_buffer(2);
+	uint32 *value = (uint32*)plist_get_buffer(2);
 	
 	//TODO: Keep in-memory queue, lazily write to flash
 	push_sensor_value( mib_addr, metadata, value );
